requirements.txt
django
djangorestframework
beautifulsoup4
requests
coreapi
django-filter